feat: Enhance ph_agent integration by adding chat session functionality for CF Portfolio and CF Security
