+++ cognitive_folio/patches.txt
@@ -4,3 +4,4 @@
 
 [post_model_sync]
 # Patches added in this section will be executed after doctypes are migrated
+cognitive_folio.patches.v16_5_0.seed_ph_agent_integration
